refactor: move utils into another file
